Move DateTimeWriteError to the pattern module
