+++ components/datetime/src/lib.rs
@@ -106,7 +106,7 @@ pub(crate) mod raw;
 pub mod scaffold;
 pub(crate) mod size_test_macro;
 
-pub use error::{DateTimeFormatterLoadError, DateTimeWriteError, MismatchedCalendarError};
+pub use error::{DateTimeFormatterLoadError, MismatchedCalendarError};
 
 pub use format::DateTimeInputUnchecked;
 pub use neo::DateTimeFormatter;
